Add org reference to events
Add a new `org` field to the events schema. This references the id from orgs
schema, which happens to be name of the folder under `content/orgs`.

Retrieving the reference is pretty trivial in EventPost - we just need a
`getEntry` method.

This is a little tricker in the index page however, since references defined in
the schema must be queried separately. For now, use a naive approach of querying
the entire org collection, then match by `org.id === event.data.org.id`. It's
not elegant, but it works.