ops: modify preview workflow to run on manual dispatch
Currenly, PRs from forked repos will fail CI, because workflows triggered from
forked repos do not have access to repository secrets.

Fix this by changing the preview workflow to be run via workflow dispatch (the
instructions are linked in a comment).

Consequently, we can no longer expect that PRs automatically trigger a
deployment to the preview environment - we have to do that manually. This is
probably a better approach anyway, since we will gain finer control to which
branch we are interested in previewing, instead of having to guess, especially
when we have multiple active PRs.